Modif d'utilisateur

+++ src/app/users/user-page/user-card/user-card.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { Router } from '@angular/router';
 import { User } from 'src/app/shared/models/user';
 import { UserService } from 'src/app/shared/services/user.service';
@@ -11,6 +11,8 @@ import { UserService } from 'src/app/shared/services/user.service';
 export class UserCardComponent implements OnInit {
 
   @Input() user!: User;
+  // Output: remonté d'information vers le composant parent
+  @Output() deleteEmitter: EventEmitter<string> = new EventEmitter<string>();
 
   constructor(private userService: UserService, private router: Router) { }
 
@@ -18,14 +20,17 @@ export class UserCardComponent implements OnInit {
   }
 
   deleteUser(): void {
-    this.userService.deleteUserById(this.user.id)
-      .subscribe((_)=>{
-        this.router.navigate(['../status'])
-      })
+    if (this.user.id) {
+      this.userService.deleteUserById(this.user.id)
+        .subscribe((_) => {
+          // this.router.navigate(['../status'])
+          this.deleteEmitter.emit();
+        })
+    }
   }
 
   editUser(): void {
-
+    this.router.navigate([`../users/${this.user.id}/edit`])
   }
 
 }
